refactor: add Agent interface for orchestrator and planner
Introduce a base Agent interface (`src/agent.ts`) that standardizes
agent creation following the same pattern used for providers and issue
fetchers: interface + boot function + registry.

- `Agent` base interface with `name` and `cleanup()` lifecycle
- `PlannerAgent` extends Agent with `plan()` method
- `OrchestratorAgent` extends Agent with `orchestrate()` method
- Agent registry (`src/agents/index.ts`) with `AGENT_NAMES` and
  type-safe `bootPlanner`/`bootOrchestrator` exports
- CLI updated to use `bootOrchestrator()` agent pattern
- Old standalone `planner.ts` and `orchestrator.ts` removed
